Move HaloAdjustedDatasource to separate file
